% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pooling.R
\name{data_pooling}
\alias{data_pooling}
\title{Aggregate Data by Time Interval}
\usage{
data_pooling(dt, variable, x_min = NULL, by, typ_fun = NULL)
}
\arguments{
\item{dt}{A \code{data.table} containing the data to be aggregated. The \code{dt} must include a \code{TIMESTAMP} column.}

\item{variable}{The name of the variable to be aggregated. The function will apply the aggregation method specified in \code{typ_fun} or default to the pre-defined method for that variable.}

\item{x_min}{(Optional) Minimum number of non-missing data points required for the aggregation. Defaults to \code{1} if not specified.}

\item{by}{A string indicating the time unit for aggregation (e.g., "hour", "day").}

\item{typ_fun}{(Optional) A custom aggregation function to be applied. If not specified, the function uses a default aggregation function based on the variable name.}
}
\value{
A \code{data.table} containing the aggregated data, with a \code{TIMESTAMP} column and aggregated values for the specified variable.
}
\description{
This function aggregates data from a \code{data.table} based on a specified time unit and applies various aggregation functions (e.g., mean, min, max) to different variables.
}
\examples{
TIMESTAMP <- seq.POSIXt(as.POSIXct("2024-09-01 00:00:00"),
as.POSIXct("2024-10-01 01:00:00"),
by = "5 min")
TempAire_Avg <- runif(length(TIMESTAMP), 20, 50)
dt <- data.table(TIMESTAMP = TIMESTAMP, TempAire_Avg = TempAire_Avg)
data_pooling(dt, variable = "TempAire_Avg", x_min = 1, by = "hour")
}
