% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_QC.R
\name{secuencia_incorrecta}
\alias{secuencia_incorrecta}
\title{Control de Calidad de Datos de Intervalo de 5 Minutos}
\usage{
secuencia_incorrecta(dt, variable_analizada)
}
\arguments{
\item{dt}{Un \code{data.table} con una columna \code{TIMESTAMP} y una columna de valores numéricos.}

\item{variable}{El nombre de la columna en \code{dt} que contiene los valores a analizar.}
}
\value{
Un \code{data.table} corregido con intervalos de tiempo y valores ajustados.
}
\description{
Esta función está diseñada para identificar y corregir intervalos incorrectos en datos con una frecuencia de 5 minutos.
Se utiliza para procesar datos en paralelo y corregir valores faltantes basándose en los valores más cercanos.
}
\details{
\itemize{
\item La función verifica si los intervalos entre \code{TIMESTAMP} son de 5 minutos.
\item Corrige los intervalos incorrectos rellenando los valores \code{NA} con los valores más cercanos.
\item Verifica que los intervalos corregidos sean correctos y lanza un error si no lo son.
}
}
